Made Updates Per Dean
Changed by to "be" in the "trying" sentence. Combined the "Get a Snack
Item" with "Get Involved at the bottom (it was weird to open up the
list with have a snack…; added a link for the Jagtag where it was
talked about to provide a connection to online informaiton about the
Jagtag.