Updated Text
Added text referring to the sophomore bucket list in the dean's letter.
It was never reference and seemed odd to not somehow elude to it. This
change was per Becky Porter after the dean sent it to her. Also, took
the small text from the image and typed it as HTML "Your sophomore
year…" - when the email is viewed in mobile it is impossible to read
the small text since the image doesn't change.